requirements.txt
streamlit
streamlit-authenticator
pyyaml
langchain
langchain-openai
chroma
pypdf